REFACTOR: Rename parser wrapper-module
Rename to avoid namespace clashes in Windows
